requirements.txt
fastapi-asyncpg==1.0.1
SQLAlchemy==2.0.43
uvicorn==0.35.0
alembic==1.16.5
pydantic-settings==2.7.1
greenlet==3.2.4
bcrypt==4.3.0
passlib==1.7.4
email-validator==2.3.0
python-jose==3.5.0